+++ src/Components/FormContext.tsx
@@ -1,46 +1,29 @@
-import { generatePassword } from "@/Components/form"
 import { usePersistedReducer } from "@/common/hooks/usePersistedReducer"
 import { maxLengthForWords, minLengthForChars } from "@/config"
+import {
+  FormContextProps,
+  FormDispatchContextProps,
+  FormState,
+  IndexableFormValues,
+  ResultContextProps,
+  ResultState,
+} from "@/models"
+import { createCryptoKey } from "@/services/createCrypto"
 import reducer, {
   FormActionKind,
-  FormActions,
-  FormState,
   initialFormState,
   setSlidervalue,
 } from "@/services/reducers/formReducer"
-import { Dispatch, ReactNode, createContext, useCallback, useState } from "react"
-
-type FormContextProps = {
-  formState: FormState
-  generate: () => void
-  validate?: (value: number, state: FormState) => number
-}
+import { ReactNode, createContext, useCallback, useState } from "react"
 
 export const FormContext = createContext<FormContextProps>({
   formState: initialFormState,
   generate: () => undefined,
 })
 
-type FormDispatchContextProps = {
-  dispatch: Dispatch<FormActions>
-}
-
 export const FormDispatchContext = createContext<FormDispatchContextProps>({
   dispatch: () => undefined,
 })
-
-/**
- * isEdited is to differentiate edited (user-inputted) passwords from generated ones
- */
-type ResultState = {
-  passwordValue?: string
-  isEdited: boolean
-}
-
-type ResultContextProps = {
-  finalPassword: ResultState
-  setFinalPassword: Dispatch<React.SetStateAction<ResultState>>
-}
 
 export const ResultContext = createContext<ResultContextProps>({
   finalPassword: { isEdited: false },
@@ -60,10 +43,8 @@ export const FormProvider = ({ children }: { children: ReactNode }) => {
   }
 
   const validate = useCallback(
-    (value: number, state: FormState): number => {
-      const sliderValue = value
+    (sliderValue: number, state: FormState): number => {
       const { selected } = state.formValues.words
-
       if (
         selected &&
         (sliderValue > maxLengthForWords || sliderValue < 1) // should return false
@@ -79,6 +60,10 @@ export const FormProvider = ({ children }: { children: ReactNode }) => {
     [formState.formValues],
   )
 
+  function generatePassword(formValues: IndexableFormValues, sliderValue: number) {
+    return createCryptoKey(sliderValue.toString(), formValues)
+  }
+
   const generate = useCallback(() => {
     inputFieldShouldDisable()
       ? dispatch({ type: SET_DISABLED, payload: true })
@@ -89,8 +74,10 @@ export const FormProvider = ({ children }: { children: ReactNode }) => {
         validate(formState.sliderValue, formState),
       )
       setFinalPassword({ passwordValue: password, isEdited: false })
-    } catch (err) {
-      console.error(err)
+    } catch (error) {
+      if (error instanceof Error) {
+        throw new Error(error.message)
+      }
     }
   }, [formState.formValues, formState.sliderValue])
 
@@ -108,3 +95,4 @@ export const FormProvider = ({ children }: { children: ReactNode }) => {
     </FormContext.Provider>
   )
 }
+
