+++ src/models/index.ts
@@ -1,3 +1,6 @@
+import { FormActions } from "@/services/reducers/formReducer"
+import { Dispatch } from "react"
+
 export type CheckboxLabels = "Isot Kirjaimet" | "Välimerkit" | "Numerot" | "Käytä sanoja"
 
 export type InputType = "checkbox" | "input" | "radio"
@@ -17,3 +20,34 @@ export type IndexableFormValues = {
   numbers: InputValue
   randomChars: InputValue
 }
+
+export type FormState = {
+  formValues: IndexableFormValues
+  sliderValue: number
+  isDisabled: boolean
+  isEditing: boolean
+}
+
+export type FormContextProps = {
+  formState: FormState
+  generate: () => void
+  validate?: (value: number, state: FormState) => number
+}
+
+export type FormDispatchContextProps = {
+  dispatch: Dispatch<FormActions>
+}
+
+/**
+ * isEdited is to differentiate edited (user-inputted) passwords from generated ones
+ */
+export type ResultState = {
+  passwordValue?: string
+  isEdited: boolean
+}
+
+export type ResultContextProps = {
+  finalPassword: ResultState
+  setFinalPassword: Dispatch<React.SetStateAction<ResultState>>
+}
+
